crewai-most-simple-example/pyproject.toml
[project]
name = "crewai-most-simple-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.79.4",
    "crewai-tools>=0.14.0",
    "python-dotenv>=1.0.1",
]
